% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_areas.R
\name{area_variables}
\alias{area_variables}
\title{Show area variable names}
\usage{
area_variables()
}
\value{
A character vector of area variable names from the area lookup.
}
\description{
Display the variable names of areas. Useful for checking what variables can
be used for subsetting requests.
}
\examples{

# List the areas
area_variables()

}
