% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_records.R
\name{calculate_riskratio}
\alias{calculate_riskratio}
\title{Calculate risk ratio}
\usage{
calculate_riskratio(
  data,
  ethnicity_definition = "self",
  collapse_ethnicity = T,
  comparison = c("white", "black"),
  period = 12
)
}
\arguments{
\item{data}{A tibble of stop data acquired using policedatR. The first column
must be the area code variable of the geography of interest (e.g. 'lad22cd').}

\item{ethnicity_definition}{String specifying which ethnicity definition to use
for counts. 'self' has the possibility of using the 18 disaggregated categories but is likely
to have more NAs than 'officer'. 'officer' is only the 5 aggregated ethnicity categories and
will likely have fewer NAs (technically it should have 0 but in practice this isn't the case)}

\item{collapse_ethnicity}{If \code{ethnicity_definition == 'self'}, this controls
whether to use the 18 disaggregated categories or to aggregate to the 5 broader
categories. If \code{ethnicity_definition == 'self'}, \code{collapse_ethnicity} is always TRUE.}

\item{comparison}{A character vector with length 2, where the first element is
the reference level of ethnicity against which the second ('test') level is compared.
If NULL, user will be prompted to pick the reference and test ethnicities.}

\item{period}{Numeric value specifying the number of months each time period
should be.}
}
\value{
A tibble with number of stops and stop rate per 1000 population for each
ethnicity, and risk ratio and confidence interval within each area-period combination.
}
\description{
Calculate risk ratio between the stop rates of two different ethnicities. Runs
policedatR::analyse_anything() with the grouping \verb{(area, period, ethnicity)} to
produce counts for each area-period-ethnicity combination, then runs\code{epitools::riskratio()}
on the output to calculate the risk ratioand confidence interval of stop rates of specified
ethnicities for each area and period.
}
