% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_records.R
\name{count_stops}
\alias{count_stops}
\title{Count stops}
\usage{
count_stops(
  data,
  ethnicity_definition = c("self", "officer"),
  collapse_ethnicity = TRUE,
  comparison = NULL,
  period = 12
)
}
\arguments{
\item{data}{A data frame extracted using policedatR where each row is a unique
stop and the first column is the area code variable for which stops were acquired
(e.g. lad22cd for Local Authority Districts).}

\item{ethnicity_definition}{String specifying which ethnicity definition to use
for counts. 'self' has the possibility of using the 18 disaggregated categories but is likely
to have more NAs than 'officer'. 'officer' is only the 5 aggregated ethnicity categories and
will likely have fewer NAs (technically it should have 0 but in practice this isn't the case)}

\item{collapse_ethnicity}{If \code{ethnicity_definition == 'self'}, this controls
whether to use the 18 disaggregated categories or to aggregate to the 5 broader
categories. If \code{ethnicity_definition == 'self'}, \code{collapse_ethnicity} is always TRUE.}

\item{comparison}{User can optionally choose to compare between two ethnicities. This is
a character vector with length 2, where the first element is the baseline against which the
second element is compared. If this is specified, counts, rates, and incidence rate ratios will
be returned for only the two ethnicities specified.}

\item{period}{Numeric value specifying the number of months each time period
should be.}
}
\value{
If \code{comparison} is NULL, a long tibble where each row is a unique
area-period-ethnicity combination across all ethnicities. Computed statistics are
count and rate per 1000 population.

If \code{comparison} is specified, a wide tibble where each row is a unique
area-period combination across only the ethnicities specified. Computed statistics are
count, rate per 1000 population, incidence rate ratio, lower and upper confidence
interval, and p-value.
}
\description{
Count the number of stops for combinations of areas, time periods, and ethnicities.
If \code{comparison} is NULL, returns a long tibble where each row is a unique
area-period-ethnicity combination across all ethnicities. Computed statistics are
count and rate per 1000 population. If \code{comparison} is not NULL, performs a
generalised linear regression using Poisson distribution and log link function
to produce an incidence rate ratio describing the extent of the different between
the stop rates of the two ethnicities specified. In this case, it returns a wide tibble
where each row is a unique area-period combination across only the ethnicities
specified. Computed statistics are count, rate per 1000 population, incidence rate ratio,
lower and upper confidence interval values, and p-value.
}
\examples{

# Count by self-defined ethnicity in one-month periods
counted_data <- count_stops(data,
                            ethnicity_definition = "self",
                            period = 1
                           )
# Count by officer-defined ethnicity and calculate the disparity in stop rates
between White and Black people, in one-month periods.
counted_data2 <- count_stops(data,
                             ethnicity_definition = "officer",
                             period = 1,
                             comparison = c("white","black")
}
