% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fetch_functions.R
\name{fetch_geometry_data}
\alias{fetch_geometry_data}
\title{Fetch geometry data}
\usage{
fetch_geometry_data(
  base_url,
  where_clause,
  result_offset = 0,
  max_records = 2000
)
}
\arguments{
\item{base_url}{The API endpoint. String.}

\item{where_clause}{Subset of areas for which to acquire geometries. String}

\item{result_offset}{For pagination user can start acquisition from an offset record number.}

\item{max_records}{For pagination user can acquire a specific number of records. Geoportal max is 2000 but
sometimes it prefers smaller payloads.}
}
\value{
An http response object
}
\description{
Fetch geometry data workhorse for get_geometry functions. Submits GET requests to API
at geoportal.statistics.gov.uk. Can be memoised (and is in the get_geometry functions) to speed up repeat calls.
}
\details{
In practice the user does not need to directly call this function.
}
\examples{
\dontrun{
# Don't run as takes a while
# Simplified example where the polygon is defined by only three points
response <- fetch_geometry_data(
  base_url = "https://services1.arcgis.com/ESMARspQHYMw9BZ9/arcgis/rest/services/Regions_December_2022_Boundaries_EN_BFC_V2/FeatureServer/0/query",
  where_clause = "1=1",
  result_offset = 0,
  max_records = 2000
  )
}
}
