% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fetch_functions.R
\name{fetch_police_data}
\alias{fetch_police_data}
\title{Fetch police data}
\usage{
fetch_police_data(body, wait_time = 5, max_tries = 10)
}
\arguments{
\item{body}{Body of POST request. A named list consisting of two elements; a
'poly' element defining a polygon as a character string in format 'lat1,long1:lat2,long2'
and a 'date' defining a date as character string in format 'yyyy-mm'.}

\item{wait_time}{On error, the time to wait before retrying. Default is 5.}

\item{max_tries}{The number of tries before giving up the call. Default is 10.}
}
\value{
An http response object
}
\description{
Fetch police data workhorse for get_data functions. Submits POST requests to API
at data.police.uk. Can be memoised (and is in the get_data functions) to speed up repeat calls.
}
\details{
In practice the user does not need to directly call this function.
}
\examples{

# Simplified example where the polgyon is defined by only three points
fetch_police_data(body = list('poly' = "51.6094409248573,-0.127876986589682:51.6093817289251,-0.127378230001641:51.6093739825532,-0.12644271332808",
                            'date' = "2025-02"))

}
