% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fetch_functions.R
\name{fetch_population_estimates}
\alias{fetch_population_estimates}
\title{Fetch population estimates}
\usage{
fetch_population_estimates(data)
}
\arguments{
\item{data}{}
}
\value{
An S3 response object containing the response of the request.
}
\description{
API workhorse for \code{get_population_estimates()}. Communicates with NOMIS API to
get population estimates by ethnicity for local areas specified by the smallest
level of geography in data (which must be column 1). Population estimates are
from Census 2021 table TS021.
}
\details{
Can be (and is in \code{get_population_estimates()}) memoised for quicker repeat
requests.
}
\examples{

# Mock df
df <- data.frame("lad22cd" = c("E09000001","E09000002","E09000003"),
                 "x" = seq(1:3))

response <- fetch_population_estimates(df)

}
