% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_functions.R
\name{flush_cache}
\alias{flush_cache}
\title{Flush the cache}
\usage{
flush_cache()
}
\value{
Nothing. Simply deletes all files found in cache directory.
}
\description{
Delete all files in the cache directory. This may be helpful if \verb{get_*} functions
are returning unexpected results. This may happen, for example, where requests
have failed but their responses were cached, and subsequent requests are
reloading this failed request.
}
\details{
Ideally this should never be necessary as failed requests should be handled by
the \verb{get_*} functions. However, there have been a few instances during development
where it was necessary to flush the cache, so this function is provided as a
'just in case you need it' function.

Cache directory is specified by \code{rappdirs::user_cache_dir(appname = NULL, appauthor = "policedatR")}
}
\examples{

\dontrun{
# This is interactive so don't run
flush_cache()
}
}
