% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometries.R
\name{get_lad_geometries}
\alias{get_lad_geometries}
\title{Get Local Authority District geometries}
\usage{
get_lad_geometries(subset = NULL)
}
\arguments{
\item{subset}{A named list defining the areas to which to subset. Names
correspond to the variable on which to subset. Values correspond to the desired
values of the variable.}
}
\value{
A tibble of geometries where each row is local authority district
}
\description{
Acquire geometry data from geoportal.gov.uk for Local Authority Districts as
at December 2022.
}
\examples{

lad_geometries <- get_lad_geometries()

# Get just Haringey and Lambeth geometries using 'lad22nm' variable
subset_geometries <- get_lad_geometries(subset = list("lad22nm" = c("Haringey", "Lambeth")))

# Get just Haringey and Lambeth geometries using 'lad22cd' variable
subset_geometries <- get_lad_geometries(subset = list("lad22cd" = c("E09000014", "E09000022")))

}
