% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometries.R
\name{get_lsoa_geometries}
\alias{get_lsoa_geometries}
\title{Get Lower layer Super Output Area geometries}
\usage{
get_lsoa_geometries(subset = NULL)
}
\arguments{
\item{subset}{A named list defining the areas for which to acquire geometries. Names
correspond to the area variable on which to subset (run policedatR::area_variables() to see options for areas on which to subset).
Values correspond to the desired values of the area variable. Values can be single
strings or character vectors.

If NULL, will acquire all LSOA geometries and will take around 5 minutes pre-caching.}
}
\value{
An sf data frame where each row is a LSOA, with columns lsoa21cd, lsoa21nm,
shape_area and geometry.
}
\description{
Acquire geometry data from geoportal.gov.uk for Lower layer Super Output Area (LSOAs) as
at December 2021.
}
\examples{

\dontrun{
# Get all LSOAs
# Don't run as will take a long time
all_lsoas <- get_lsoa_geometries()

# Get LSOAs in the Greater London region
# Don't run as will take a while
london_lsoas <- get_lsoa_geometries(subset = list("rgn22nm" = "London"))
}
# Get LSOAs just in Haringey and Waltham Forest
haringey_waltham_lsoas <- get_lsoa_geometries(subset = list("lad22nm" = c("Haringey", "Waltham Forest")))


}
