% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometries.R
\name{get_msoa_geometries}
\alias{get_msoa_geometries}
\title{Get Middle layer Super Output Area geometries}
\usage{
get_msoa_geometries(subset = NULL)
}
\arguments{
\item{subset}{A named list defining the areas for which to acquire geometries. Names
correspond to the variable on which to subset (run policedatR::area_variables() to see options for areas on which to subset).
Values correspond to the desired values of the variable. Values can be single
strings or character vectors.

If NULL, will acquire all MSOA geometries and will take around 5 minutes.}
}
\value{
An sf data frame where each row is a MSOA, with columns msoa21cd, msoa21nm,
shape_area and geometry.
}
\description{
Acquire geometry data from geoportal.gov.uk for Middle layer Super Output Area (MSOAs) as
at December 2021.
}
\examples{

# Get all MSOAs
all_msoas <- get_msoa_geometries()

# Get MSOAs in the Greater London region
london_msoas <- get_msoa_geometries(subset = list("rgn22nm" = "London"))

# Get MSOAs just in Haringey and Waltham Forest
haringey_waltham_msoas <- get_msoa_geometries(subset = list("lad22nm" = c("Haringey", "Waltham Forest")))

}
