% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometries.R
\name{get_oa_geometries}
\alias{get_oa_geometries}
\title{Get Output Area geometries}
\usage{
get_oa_geometries(subset = NULL)
}
\arguments{
\item{subset}{A named list defining the areas for which to acquire geometries. Names
correspond to the area variable on which to subset (run policedatR::area_variables() to see options for areas on which to subset).
Values correspond to the desired values of the area variable. Values can be single
strings or character vectors.

If NULL, will acquire all OA geometries and will take around 5 minutes.}
}
\value{
An sf data frame where each row is an  OA, with columns oa21cd, shape_area and geometry.
}
\description{
Acquire geometry data from geoportal.gov.uk for Output Area (OAs) as
at December 2021.
}
\examples{

\dontrun{
# Get all OAs
# Don't run as will take a long time
all_oas <- get_oa_geometries()

# Get OAs in the Greater London region
# Don't run as will take a long time
london_oas <- get_oa_geometries(subset = list("rgn22nm" = "London"))
}

# Get OAs just in Haringey and Waltham Forest
haringey_oas <- get_oa_geometries(subset = list("lad22nm" = "Haringey"))


}
