% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometries.R
\name{get_pfa_geometries}
\alias{get_pfa_geometries}
\title{Get Police Force Area geometries}
\usage{
get_pfa_geometries(subset = NULL)
}
\arguments{
\item{subset}{A named list defining the areas for which to acquire geometries. Names
correspond to the area variable on which to subset (run policedatR::area_variables() to see options for areas on which to subset).
Values correspond to the desired values of the area variable. Values can be single
strings or character vectors.

If the areas specified are at a higher level of geography, all PFAs within the higher geographies will be returned.
If the areas specified are at a lower level of geography, the PFAs to which the lower geography belongs will be returned.

If NULL, will acquire all PFA geometries and will take around 1 minute pre-caching.}
}
\value{
An sf data frame where each row is a Local Authority District, with columns lad22cd, lad22nm,
shape_area and geometry.
}
\description{
Acquire geometry data from geoportal.gov.uk for Police Force Areas (PFAs) as
at December 2022.
}
\examples{

\dontrun{
# Get all PFA geometries
# Don't run as takes ~49 seconds
pfa_geometries <- get_pfa_geometries()
}

# Get the PFA that Haringey is a part of
subset_geometries <- get_lad_geometries(subset = list("lad22nm" = "Haringey"))


}
