% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_estimates.R
\name{get_population_estimates}
\alias{get_population_estimates}
\title{Get population estimates}
\usage{
get_population_estimates(data)
}
\arguments{
\item{data}{A data frame or tibble of data. This will usually be be an output
of another policedatR function like get_data or analyse_data. The first column
must be the geography variable for which population estimates are desired.}
}
\value{
A long-format tibble with poplulation estimates for each ethnicity
for each (lowest-level) geography in data.
}
\description{
Acquire population estimates broken down by ethnicity for the unique areas
in data via the NOMIS API. Uses an internal fetch function taht communicates
with the NOMIS API.
}
\examples{

# Mock df
df <- data.frame("lad22cd" = c("E09000001","E09000002","E09000003"),
                 "x" = seq(1:3))

population_estimates <- get_population_estimates(df)

}
