% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometries.R
\name{get_region_geometries}
\alias{get_region_geometries}
\title{Get Region geometries}
\usage{
get_region_geometries(subset = NULL)
}
\arguments{
\item{subset}{A named list defining the areas for which to acquire geometries. Names
correspond to the area variable on which to subset (run policedatR::area_variables() to see options for areas on which to subset).
Values correspond to the desired values of the area variable. Values can be single
strings or character vectors.

If the areas specified are at a lower level of geography, the Regions to which the lower geography belongs will be returned.

If NULL, will acquire all Region geometries and will take around 1 minute pre-caching.}
}
\value{
An sf data frame where each row is a Region, with columns rgn22cd, rgn22nm,
shape_area and geometry.
}
\description{
Acquire geometry data from geoportal.gov.uk for Regions as
at December 2022.
}
\examples{

\dontrun{
# Get all region geometries
# Don't run as takes ~12 seconds
region_geometries <- get_region_geometries()

}
# Get the Region that Haringey is a part of
subset_geometries <- get_region_geometries(subset = list("lad22nm" = "Haringey"))


}
