% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_areas.R
\name{list_areas}
\alias{list_areas}
\title{List available areas}
\usage{
list_areas(area_type = "lad")
}
\arguments{
\item{area_type}{A string definiting the area type for which distinct values
are sought. Options are "oa" (Output Area), "lsoa", (Lower layer Super Output
area), "msoa" (Middle layer Super Output Area), and "lad" (Local Authority
District)}
}
\value{
A character vector of the unique values of area_type.
}
\description{
List unique values of the different area types that could be used to get
data from the Police API.
}
\examples{

list_areas(area_type = "lad")
}
