% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/police_data_recency.R
\name{newest_data}
\alias{newest_data}
\title{Retrieve most recent date for which data in API are available}
\usage{
newest_data(timeout = 10)
}
\value{
A string describing date in yyyy-mm format
}
\description{
Retrieve most recent date for which data in API are available
}
\examples{

newest_data()

}
