% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_helpers.R
\name{show_analysis_variables}
\alias{show_analysis_variables}
\title{Show analysis variables}
\usage{
show_analysis_variables()
}
\value{
A tibble where \code{key} is the value to be input into \code{analysis_variables}
and \code{description} is a description of what it means
}
\description{
Show what values can be used in \code{analysis_variables} argument of analysis
functions and a description of what each value means.
}
\examples{

policedatR::show_analysis_variables()

}
