% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometries.R
\name{get_lad_geometries}
\alias{get_lad_geometries}
\title{Get Local Authority District geometries}
\usage{
get_lad_geometries(subset = NULL)
}
\arguments{
\item{subset}{A named list defining the areas for which to acquire geometries. Names
correspond to the area variable on which to subset (run policedatR::area_variables() to see options for areas on which to subset).
Values correspond to the desired values of the area variable. Values can be single
strings or character vectors.

If NULL, will acquire all LAD geometries and will take around 1 minute pre-caching.}
}
\value{
An sf data frame where each row is a Local Authority District, with columns lad22cd, lad22nm,
shape_area and geometry.
}
\description{
Acquire geometry data from geoportal.gov.uk for Local Authority Districts as
at December 2022.
}
\examples{

lad_geometries <- get_lad_geometries()

# Get just Haringey and Lambeth geometries using 'lad22nm' variable
subset_geometries <- get_lad_geometries(subset = list("lad22nm" = c("Haringey", "Lambeth")))

# Get just Haringey and Lambeth geometries using 'lad22cd' variable
subset_geometries <- get_lad_geometries(subset = list("lad22cd" = c("E09000014", "E09000022")))

}
